Move the function-definition to its own file
